MC101-2005/Task2/pyproject.toml
[project]
name = "d2"
version = "0.1.0"
description = ""
authors = [
    {name = "siamak",email = "khatami.sia@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[all] (>=0.117.1,<0.118.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "argon2-cffi (>=23.1.0,<24.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
